Move codel implementation to header file
